product form and category form

+++ pages/api/carts/[id].ts
@@ -8,9 +8,13 @@ const getItem = async (req: IRequest, res: NextApiResponse) => {
 	try {
 		const user = req.user;
 		const id = req.query.id as string;
-		let item = await db.collection(`users/${user.id}/cart`).doc(id).get();
+		let item = await db.collection('carts').doc(id).get();
 		if (!item.exists)
 			return res.status(400).json({error: "Item Didn't Exists!"});
+		
+		if(item.get('user') !== user.uid)
+			return res.status(403).json({error: "This item does not belongs to you!"});
+		
 		return res.status(200).json({id: item.id, ...item.data()});
 	} catch (err) {
 		return res.status(400).json({error: (err as Error).message});
@@ -22,6 +26,15 @@ const updateItem = async (req: IRequest, res: NextApiResponse) => {
 		const id = req.query.id as string;
 		const body = req.body as Partial<ICartItem>;
 		const user = req.user;
+
+		let itemRef = db.collection('carts').doc(id);
+		let item = await itemRef.get();
+		if (!item.exists)
+			return res.status(400).json({error: 'Item no longer exists!'});
+
+		if(item.get('user') !== user.uid)
+			return res.status(403).json({error: 'This item does not belongs to you!'});
+		
 		if (
 			body.quantity == null ||
 			(body.quantity != null && typeof body.quantity !== 'number')
@@ -30,17 +43,13 @@ const updateItem = async (req: IRequest, res: NextApiResponse) => {
 				.status(400)
 				.json({error: 'Quantity is required and should be a positive number!'});
 
-		let itemRef = db.collection(`users/${user.id}/cart`).doc(id);
-		let item = await itemRef.get();
-		if (!item.exists)
-			return res.status(400).json({error: 'Item no longer exists!'});
-
 		let product = await (item.get('product') as DocumentReference).get();
 
 		let updatedData = {
 			quantity: body.quantity,
 			price: product.get('price') * body.quantity,
 		};
+
 		await itemRef.update(updatedData);
 		return res.status(200).json(updatedData);
 	} catch (err) {
@@ -52,7 +61,12 @@ const deleteItem = async (req: IRequest, res: NextApiResponse) => {
 	try {
 		const user = req.user;
 		const id = req.query.id as string;
-		let itemRef = db.collection(`users/${user.id}/cart`).doc(id);
+		let itemRef = db.collection('carts').doc(id);
+		let item = await itemRef.get();
+
+		if(item.get('user') !== user.uid)
+			return res.status(403).json({error: 'This item does not belongs to you!'});
+
 		await itemRef.delete();
 		return res.status(204);
 	} catch (err) {
